+++ JtProject/src/main/java/ca/concordia/eats/ConcordiaEatsApplication.java
@@ -1,13 +1,13 @@
-package com.jtspringproject.JtSpringProject;
+package ca.concordia.eats;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class JtSpringProjectApplication {
+public class ConcordiaEatsApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(JtSpringProjectApplication.class, args);
+		SpringApplication.run(ConcordiaEatsApplication.class, args);
 	}
 
 }
