+++ JtProject/src/main/java/ca/concordia/eats/controller/MainController.java
@@ -1,0 +1,8 @@
+package ca.concordia.eats.controller;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class MainController {
+
+}
