+++ JtProject/src/main/java/ca/concordia/eats/controller/OrderController.java
@@ -1,0 +1,7 @@
+package ca.concordia.eats.controller;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class OrderController {
+}
