+++ JtProject/src/main/java/ca/concordia/eats/dao/OrderDaoImpl.java
@@ -1,0 +1,7 @@
+package ca.concordia.eats.dao;
+
+import org.springframework.stereotype.Repository;
+
+@Repository
+public class OrderDaoImpl implements UserDao {
+}
