+++ JtProject/src/main/java/ca/concordia/eats/dao/PromotionDao.java
@@ -1,0 +1,4 @@
+package ca.concordia.eats.dao;
+
+public interface PromotionDao {
+}
