+++ JtProject/src/main/java/ca/concordia/eats/dao/UserDaoImpl.java
@@ -1,0 +1,7 @@
+package ca.concordia.eats.dao;
+
+import org.springframework.stereotype.Repository;
+
+@Repository
+public class UserDaoImpl implements UserDao {
+}
