+++ JtProject/src/main/java/ca/concordia/eats/dto/Admin.java
@@ -1,0 +1,4 @@
+package ca.concordia.eats.dto;
+
+public class Admin extends User{
+}
