+++ JtProject/src/main/java/ca/concordia/eats/dto/BasketItem.java
@@ -1,0 +1,4 @@
+package ca.concordia.eats.dto;
+
+public class BasketItem {
+}
