+++ JtProject/src/main/java/ca/concordia/eats/dto/Category.java
@@ -1,0 +1,31 @@
+package ca.concordia.eats.dto;
+
+public class Category {
+
+    private int id;
+    private String name;
+
+    public Category() {
+    }
+
+    public Category(int id, String name) {
+        this.id = id;
+        this.name = name;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
