+++ JtProject/src/main/java/ca/concordia/eats/dto/Promotion.java
@@ -1,0 +1,4 @@
+package ca.concordia.eats.dto;
+
+public class Promotion {
+}
