+++ JtProject/src/main/java/ca/concordia/eats/service/OrderServiceImpl.java
@@ -1,0 +1,7 @@
+package ca.concordia.eats.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class OrderServiceImpl implements OrderService {
+}
