+++ JtProject/src/main/java/ca/concordia/eats/service/ProductService.java
@@ -1,0 +1,24 @@
+package ca.concordia.eats.service;
+
+import ca.concordia.eats.dto.Category;
+import ca.concordia.eats.dto.Product;
+
+import java.util.List;
+
+public interface ProductService {
+
+    // CRUD PRODUCT
+    public List<Product> fetchAllProducts();
+    public Product fetchProductById(int productId);
+    public Product createProduct(Product product);
+    public Product updateProduct(Product product);
+    public boolean removeProductById(int productId);
+
+    // CRUD CATEGORY
+    public List<Category> fetchAllCategories();
+    public Category fetchCategoryById(int categoryId);
+    public Category createCategory(Category category);
+    public Category updateCategory(Category category);
+    public boolean removeCategoryById(int categoryId);
+
+}
