+++ JtProject/src/main/java/ca/concordia/eats/service/PromotionService.java
@@ -1,0 +1,4 @@
+package ca.concordia.eats.service;
+
+public interface PromotionService {
+}
