+++ JtProject/src/main/java/ca/concordia/eats/service/UserService.java
@@ -1,0 +1,4 @@
+package ca.concordia.eats.service;
+
+public interface UserService {
+}
